PKI-APP/requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.3
fastapi==0.103.0
h11==0.14.0
idna==3.4
jwt==1.3.1
pika==1.3.2
psycopg2==2.9.7
py-hsm==2.5.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
PyKCS11==1.5.12
pyOpenSSL==23.2.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.23.2